fix: update Docker build for amd64 and enhance Cloudflare integration (#22)
Focus on targeting the amd64 architecture in Docker builds simplifies
the process by removing multi-platform support. The changes also include
updates to the Cloudflare API token handling, ensuring consistent naming
across the codebase, and implementing a migration service for database
synchronization. Additional enhancements include the introduction of a
Metric model for metrics storage and various improvements to the
authentication and Caddy service configurations.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: rhad00 <radu.malica@gmail.com>